pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "duckdb>=1.4.3",
    "numpy>=2.3.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.52.1",
    "streamlit-aggrid>=1.2.1",
]
